Adjustments to file ktx-headings
Large parts of the comments rewritten (and compacted to the top of the
file, now the code is one long block). Made 'options' available for the
generation file, i.e. headings can be generated without the possibility
to choose the style. Also added 'driver' to the meta part of the file,
i.e. a documentation can be generated separately from the main one.