Major changes in documentation of bibliography
Added info about the 'doilink' field (also how it is applied. Added
documentation about the exact implementation of 'collaboration' as an
additional field for articles.
Besides, resorted some code of the file and completely removed the
obsolete option 'overwrite' (was used for overwriting Biblatex
standards). Since Biblatex is now included in this package, it is
redundant.