Fixed bug when trying to remove eprint links
With the class option 'print', the class is supposed to check whether a
journal title for an bibliography entry of type 'article' is
provided. If yes, the eprint should not be shown in the
bibliography. This guarantees to only have eprint links when no other
source can be provided (e.g. when submitted, but not yet published).